split common impl
