+++ crates/av/src/audio_source_wasm.rs
@@ -1,48 +1,24 @@
-use std::path::PathBuf;
-
 use bevy::{platform::collections::HashMap, prelude::*, render::view::RenderLayers};
 use common::{
-    sets::SetupSets,
-    structs::{AudioEmitter, AudioSettings, AudioType, PrimaryCameraRes, PrimaryUser, SystemAudio},
-    util::{AudioReceiver, VolumePanning},
+    structs::{AudioEmitter, AudioSettings, AudioType, PrimaryUser, SystemAudio},
+    util::VolumePanning,
 };
-use dcl::interface::ComponentPosition;
-use dcl_component::{proto_components::sdk::components::PbAudioSource, SceneComponentId};
-use ipfs::{
-    ipfs_path::{IpfsPath, IpfsType},
-    IpfsAssetServer,
+use ipfs::IpfsAssetServer;
+use scene_runner::{ContainingScene, SceneEntity};
+use web_sys::{
+    js_sys::Float32Array, AudioBuffer, AudioBufferSourceNode, AudioContext,
+    AudioScheduledSourceNode, GainNode, StereoPannerNode,
 };
-use scene_runner::{
-    renderer_context::RendererSceneContext, update_world::AddCrdtInterfaceExt, ContainingScene,
-    SceneEntity,
-};
-use web_sys::{AudioBuffer, AudioBufferSourceNode, AudioContext, AudioScheduledSourceNode, GainNode, StereoPannerNode, js_sys::Float32Array};
 
-#[derive(Component, Debug)]
-pub struct AudioSource(PbAudioSource);
+pub struct AudioSourcePluginImpl;
 
-impl From<PbAudioSource> for AudioSource {
-    fn from(value: PbAudioSource) -> Self {
-        Self(value)
-    }
-}
-
-pub struct AudioSourcePlugin;
-
-impl Plugin for AudioSourcePlugin {
+impl Plugin for AudioSourcePluginImpl {
     fn build(&self, app: &mut App) {
         // still use kira for audio source asset management
-        app.add_plugins(bevy_kira_audio::AudioPlugin);
         app.init_non_send_resource::<HtmlAudioContext>();
-        app.add_event::<SystemAudio>();
-        app.add_crdt_lww_component::<PbAudioSource, AudioSource>(
-            SceneComponentId::AUDIO_SOURCE,
-            ComponentPosition::EntityOnly,
-        );
         app.add_systems(
             PostUpdate,
             (
-                map_scene_audio_sources,
                 manage_audio_sources,
                 play_system_audio,
                 |mut ctx: NonSendMut<HtmlAudioContext>,
@@ -54,7 +30,6 @@ impl Plugin for AudioSourcePlugin {
                 .chain()
                 .after(TransformSystem::TransformPropagate),
         );
-        app.add_systems(Startup, setup_audio.in_set(SetupSets::Main));
     }
 }
 
@@ -136,7 +111,9 @@ impl HtmlAudioContext {
             .ok()?;
 
         if let Some(offset) = offset {
-            source_node.start_with_when_and_grain_offset(0.0, offset as f64).ok()?;
+            source_node
+                .start_with_when_and_grain_offset(0.0, offset as f64)
+                .ok()?;
         } else {
             source_node.start().ok()?;
         }
@@ -151,75 +128,6 @@ impl HtmlAudioContext {
     }
 }
 
-fn setup_audio(mut commands: Commands, camera: Res<PrimaryCameraRes>) {
-    commands
-        .entity(camera.0)
-        .try_insert(AudioReceiver::default());
-}
-
-fn map_scene_audio_sources(
-    mut commands: Commands,
-    mut query: Query<
-        (
-            Entity,
-            &SceneEntity,
-            &AudioSource,
-            Option<&mut AudioEmitter>,
-        ),
-        Changed<AudioSource>,
-    >,
-    ipfas: IpfsAssetServer,
-    scenes: Query<&RendererSceneContext>,
-) {
-    for (ent, scene_ent, audio_source, maybe_emitter) in query.iter_mut() {
-        let Ok(scene) = scenes.get(scene_ent.root) else {
-            warn!("failed to load audio source scene");
-            continue;
-        };
-        let ipfs_path = PathBuf::from(&IpfsPath::new(IpfsType::new_content_file(
-            scene.hash.to_owned(),
-            audio_source.0.audio_clip_url.to_owned(),
-        )));
-
-        let handle = maybe_emitter
-            .and_then(|mut existing| {
-                if existing.handle.path().is_none_or(|p| p.path() != ipfs_path) {
-                    None
-                } else {
-                    Some(std::mem::take(&mut existing.handle))
-                }
-            })
-            .unwrap_or_else(|| ipfas.asset_server().load(ipfs_path));
-
-        let seek_time = audio_source.0.current_time.and_then(|time| {
-            if time < 1e6 {
-                Some(time)
-            } else {
-                warn!(
-                    "ignoring ridiculous time offset {} for audio clip `{}`",
-                    time, audio_source.0.audio_clip_url
-                );
-                None
-            }
-        });
-
-        let emitter = AudioEmitter {
-            handle,
-            playing: audio_source.0.playing(),
-            playback_speed: audio_source.0.pitch.unwrap_or(1.0),
-            r#loop: audio_source.0.r#loop(),
-            volume: audio_source.0.volume.unwrap_or(1.0),
-            global: audio_source.0.global(),
-            seek_time,
-            ty: AudioType::Scene,
-        };
-
-        info!("emitter: {emitter:?}");
-
-        commands.entity(ent).try_insert(emitter);
-    }
-}
-
 pub struct AudioGraphHtmlElements {
     pub source_node: AudioBufferSourceNode,
     pub gain_node: GainNode,
@@ -230,7 +138,10 @@ pub struct AudioGraphHtmlElements {
 
 impl AudioGraphHtmlElements {
     pub fn stop(&self, now: f64) {
-        let _ = self.gain_node.gain().linear_ramp_to_value_at_time(0.0, now + 0.01);
+        let _ = self
+            .gain_node
+            .gain()
+            .linear_ramp_to_value_at_time(0.0, now + 0.01);
         let node: &AudioScheduledSourceNode = self.source_node.as_ref();
         let _ = node.stop_with_when(now + 0.01);
     }
